src/sample2/scraping/requirements.txt
google-cloud==0.34.0
jpholiday==0.1.4
requests==2.24.0
requests-html==0.10.0